+++ tests/conftest.py
@@ -1,12 +1,14 @@
 # pytest configuration file
 
-
 import os
-import ispyb.sqlalchemy
 import pytest
+from sqlalchemy import create_engine
+from sqlalchemy.orm import scoped_session, sessionmaker
+
+import ispyb.sqlalchemy
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def testconfig():
     """Return the path to a configuration file pointing to a test database."""
     config_file = os.path.abspath(
@@ -40,6 +42,26 @@ def testconfig_ws():
     return config_file
 
 
-@pytest.fixture
-def alchemy(testconfig):
-    return ispyb.sqlalchemy.session(testconfig)
+@pytest.fixture(scope="session")
+def db_engine(testconfig):
+    """Yields a SQLAlchemy engine"""
+    engine = create_engine(
+        ispyb.sqlalchemy.url(testconfig), connect_args={"use_pure": True}
+    )
+    yield engine
+    engine.dispose()
+
+
+@pytest.fixture(scope="session")
+def db_session_factory(db_engine):
+    """Returns a SQLAlchemy scoped session factory"""
+    return scoped_session(sessionmaker(bind=db_engine))
+
+
+@pytest.fixture(scope="function")
+def db_session(db_session_factory):
+    """Yields a SQLAlchemy connection which is rollbacked after the test"""
+    session_ = db_session_factory()
+    yield session_
+    session_.rollback()
+    session_.close()