+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 Unreleased / master
 -------------------
 
+* ``ispyb.sqlalchemy.url()`` is a function that generates the SQLAlchemy connection URL from the ISPyB configuration
+
 6.0.1 (2021-03-16)
 ------------------